api/requirements.txt
mailwizz-python-sdk